+++ frontend_v1/app/src/main/java/com/example/myapplication/RequestCallbacks/RecordingCallback.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface RecordingCallback {
 
-    public void run(Activity activity, List<short[]> data);
+    public void run(Activity activity, List<float[]> data);
 
 }
